Added link to resume.
I ended up going a bit off the main path with how I structured some stuff. I ended up thinking to much about as far as what to actually write in come of the fields and ended up just throwing some lorem in there since I had already used the type of element that was to be practiced in other places throughout the document.